added file fetcher lambda [WIP]
still need to remove the hardcoded bucket name

+++ lib/restHttp/src/main/scala/com/gu/util/zuora/RestRequestMaker.scala
@@ -77,7 +77,7 @@ object RestRequestMaker extends Logging {
       } yield ()
     }
 
-    case class DownloadStream(stream: InputStream, lengthBytes: Int)
+    case class DownloadStream(stream: InputStream, lengthBytes: Long)
     def download(path: String): DownloadStream = {
       logger.info("HELLO!!!")
       logger.info(s"path is $path")
@@ -89,7 +89,7 @@ object RestRequestMaker extends Logging {
       logger.info(s"reqsponse headers ${response.headers()}")
       DownloadStream(
         stream = response.body().byteStream,
-        lengthBytes = response.header("content-length").toInt
+        lengthBytes = response.header("content-length").toLong
       ) //todo see what to do if we don't get the lenght
     }
   }
